+++ src/packer/engine/sandbox/static/__init__.py
@@ -1,0 +1,1 @@
+from packer.engine.sandbox.static import ast_rules  # noqa: F401  (self-registration)